requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.6.1

PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.0
psutil==7.2.2

pillow==12.2.0